✅ Split example_test.go to each function
